put themes in separate directory
